requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cramjam==2.9.1
cycler==0.12.1
dashscope==1.22.2
datasets==3.5.0
decorator==4.4.2
dill==0.3.8
distro==1.9.0
fastapi==0.115.11
fastparquet==2024.11.0
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.17.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
gradio==5.20.1
gradio_client==1.7.2
groovy==0.1.2
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jiter==0.8.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
moviepy==1.0.3
multidict==6.1.0
multiprocess==0.70.16
numpy==2.2.3
openai==1.65.4
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==10.4.0
proglog==0.1.10
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
ruff==0.9.9
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.18.3
yt-dlp==2025.3.27
